.Net application that works online and offline Smart Client Concept
Introduction

In this article, we are going learn to create an application that works
online and offline, oh you might think what’s this an application online
and offline really means if the application is connected to an internet
connection then the application is online and if the application is not
connected to any internet connection it's offline.

For making this kind of application I have created 2 applications.
1) Asp.net MVC web application
2) Windows application
You again might think why 2 application right?